requirement_gcp.txt
numpy==1.19.1
pytest==7.3.1
urllib3==1.26.7
allure-pytest==2.13.2
pymysql==1.0.3
python-dotenv==0.21.1
webdriver-manager==3.8.6
selenium==4.9.0
pandas==1.3.0
openpyxl==3.1.2
pytest-xdist==3.3.1
pytest-rerunfailures==11.1.2